Fix: Se arreglaron problemas en los stored procedures y se crea la clave del empleado automaticamente

+++ BD_Scripts/procedures.sql
@@ -206,8 +206,7 @@ BEGIN
 END;
 $$ LANGUAGE plpgsql;
 
-CREATE OR REPLACE FUNCTION AddUser_Role(rol_id INTEGER,
-                                             use_id INTEGER)
+CREATE OR REPLACE FUNCTION AddUser_Role(rol_id BIGINT, use_id BIGINT)
   RETURNS INTEGER AS
 $$
 DECLARE
